♻️ Moved commands to cli sub package
