+++ qt_dev_helper/cli/commands/designer.py
